Cross validations runs, but bad stuff when NaN

+++ wikAPIdia-sr/src/main/java/org/wikapidia/sr/Evaluation/CrossValidation.java
@@ -135,7 +135,7 @@ public class CrossValidation {
 
         }
 
-        int sumError = 0;
+        double sumError = 0;
